Refactor metal compute pipeline
