+++ candle-metal-kernels/src/metal/compute_pipeline.rs
@@ -1,0 +1,24 @@
+use objc2::{rc::Retained, runtime::ProtocolObject};
+use objc2_metal::MTLComputePipelineState;
+
+#[derive(Clone, Debug)]
+pub struct ComputePipeline {
+    raw: Retained<ProtocolObject<dyn MTLComputePipelineState>>,
+}
+
+unsafe impl Send for ComputePipeline {}
+unsafe impl Sync for ComputePipeline {}
+
+impl ComputePipeline {
+    pub fn new(raw: Retained<ProtocolObject<dyn MTLComputePipelineState>>) -> ComputePipeline {
+        ComputePipeline { raw }
+    }
+
+    pub fn as_ref(&self) -> &ProtocolObject<dyn MTLComputePipelineState> {
+        &self.raw
+    }
+
+    pub fn max_total_threads_per_threadgroup(&self) -> usize {
+        self.raw.maxTotalThreadsPerThreadgroup()
+    }
+}
