+++ candle-metal-kernels/src/metal/mod.rs
@@ -1,5 +1,6 @@
 pub mod buffer;
 pub mod command_buffer;
+pub mod compute_pipeline;
 pub mod device;
 pub mod encoder;
 pub mod library;
@@ -7,6 +8,7 @@ pub mod metal_utils;
 
 pub use buffer::*;
 pub use command_buffer::*;
+pub use compute_pipeline::*;
 pub use device::*;
 pub use encoder::*;
 pub use library::*;
